Fix 1.8 support
A person using this plugin had errors in the console when triggering the PlayerInteractEvent event, I noticed that in 1.10 they added the getItemInMainHand() and it is used here. So basically you broke the 1.8 support, I added back getItemInHand() which does support the version and eliminates the error completely.

Also the isNull on the getInventory() is redundant, because already is overrided by a @NotNull.